Reorganize dynamic workspace suite files.
