+++ meteor-app/imports/ui/components/MapWithToolbar.js
