+++ meteor-app/imports/ui/pages/dynamic-workspace/index.less
@@ -82,12 +82,7 @@
     }
   }
 
-  @import './suites/dataset.tab.info.less';
-  @import './suites/dataset.tab.download.less';
-  @import './suites/dataset.tab.overlay.less';
-  @import './suites/dataset.tab.analytics.less';
-  @import './suites/dataset.tab.model.less';
-  @import './suites/dataset.tab.metadata.less';
+  @import './suites/dataset/style';
 
   .layer-list {
     .layer-list__item {
