+++ meteor-app/imports/ui/pages/dynamic-workspace/suites/SubComponentClass.js
@@ -7,7 +7,7 @@
 import _ from 'lodash';
 
 export default
-class SubComponent {
+class SubComponentClass {
   // Override this.
   static propTypes = {};
   // Override this.
