+++ meteor-app/imports/ui/pages/dynamic-workspace/suites/dataset/mapLayerLegendRenderers.js
