+++ meteor-app/imports/ui/pages/dynamic-workspace/suites/dataset/style.less
@@ -1,0 +1,6 @@
+@import './tabs/info/style';
+@import './tabs/download/style';
+@import './tabs/overlay/style';
+@import './tabs/analytics/style';
+@import './tabs/model/style';
+@import './tabs/metadata/style';
