+++ meteor-app/imports/ui/pages/dynamic-workspace/suites/dataset/tabs/download/index.js
@@ -14,7 +14,7 @@ import {
   MarkDownRenderer,
 } from '/imports/ui/helpers';
 
-import TabBaseClass from './dataset.tab.BaseClass';
+import TabBaseClass from '../BaseClass';
 
 export default
 class DownloadTab extends TabBaseClass {
