+++ meteor-app/imports/ui/pages/dynamic-workspace/suites/dataset/tabs/metadata/index.js
@@ -5,7 +5,7 @@ import {
   MarkDownRenderer,
 } from '/imports/ui/helpers';
 
-import TabBaseClass from './dataset.tab.BaseClass';
+import TabBaseClass from '../BaseClass';
 
 export default
 class MetadataTab extends TabBaseClass {
